:recycle: GET 관련 요청은 ~QueryService 에서 처리하도록 분리
