+++ src/main/java/com/umc/ttg/domain/store/application/StoreCommandService.java
@@ -10,14 +10,4 @@ public interface StoreCommandService {
 
     BaseResponseDto<StoreCreateResponseDto> saveStore(StoreCreateRequestDto storeCreateRequestDto) throws IOException;
 
-    BaseResponseDto<StoreFindResponseDto> findStore(Long storeId, Long memberId);
-
-    BaseResponseDto<HomeResponseDto> getHome(Long memberId);
-
-    BaseResponseDto<Page<StoreFindByRegionResponseDto>> findStoreByRegion(Long regionId, int page, int size, Long memberId);
-
-    BaseResponseDto<Page<StoreFindByMenuResponseDto>> findStoreByMenu(Long menuId, int page, int size, Long memberId);
-
-    BaseResponseDto<Page<StoreSearchResponseDto>> searchStore(String keyword, int page, int size, Long memberId);
-
 }
