+++ src/main/java/com/umc/ttg/domain/store/application/StoreQueryService.java
@@ -1,0 +1,19 @@
+package com.umc.ttg.domain.store.application;
+
+import com.umc.ttg.domain.store.dto.*;
+import com.umc.ttg.global.common.BaseResponseDto;
+import org.springframework.data.domain.Page;
+
+public interface StoreQueryService {
+
+    BaseResponseDto<StoreFindResponseDto> findStore(Long storeId, Long memberId);
+
+    BaseResponseDto<HomeResponseDto> getHome(Long memberId);
+
+    BaseResponseDto<Page<StoreFindByRegionResponseDto>> findStoreByRegion(Long regionId, int page, int size, Long memberId);
+
+    BaseResponseDto<Page<StoreFindByMenuResponseDto>> findStoreByMenu(Long menuId, int page, int size, Long memberId);
+
+    BaseResponseDto<Page<StoreSearchResponseDto>> searchStore(String keyword, int page, int size, Long memberId);
+
+}
